refactor(go_backend): delete legacy checker emitter path
