+++ compiler/go_backend/modules.go
@@ -9,6 +9,7 @@ import (
 	"github.com/akonwi/ard/backend"
 	"github.com/akonwi/ard/checker"
 	"github.com/akonwi/ard/frontend"
+	"github.com/akonwi/ard/go_backend/lowering"
 )
 
 func writeGeneratedProject(generatedDir string, project *checker.ProjectInfo, entrypoint checker.Module) error {
@@ -38,7 +39,11 @@ func writeGeneratedProject(generatedDir string, project *checker.ProjectInfo, en
 	}
 
 	written := map[string]struct{}{}
-	for _, mod := range requiredImportedModules(entrypoint.Program(), project.ProjectName) {
+	mods, err := requiredImportedModules(entrypoint, project.ProjectName)
+	if err != nil {
+		return err
+	}
+	for _, mod := range mods {
 		if err := writeImportedModule(generatedDir, project.ProjectName, mod, written); err != nil {
 			return err
 		}
@@ -46,18 +51,23 @@ func writeGeneratedProject(generatedDir string, project *checker.ProjectInfo, en
 	return nil
 }
 
-func requiredImportedModules(program *checker.Program, projectName string) []checker.Module {
-	if program == nil {
-		return nil
+func requiredImportedModules(module checker.Module, projectName string) ([]checker.Module, error) {
+	if module == nil || module.Program() == nil {
+		return nil, nil
 	}
-	imports := collectModuleImports(program.Statements, projectName)
-	mods := make([]checker.Module, 0, len(program.Imports))
-	for _, mod := range sortedModules(program.Imports) {
-		if _, ok := imports[moduleImportPath(projectName, mod.Path())]; ok {
-			mods = append(mods, mod)
+	irModule, err := lowering.LowerModuleToBackendIR(module, packageNameForModulePath(module.Path()), false, projectName)
+	if err != nil {
+		return nil, err
+	}
+	mods := make([]checker.Module, 0, len(irModule.ImportedModulePaths))
+	for _, path := range irModule.ImportedModulePaths {
+		mod, ok := module.Program().Imports[path]
+		if !ok {
+			continue
 		}
+		mods = append(mods, mod)
 	}
-	return mods
+	return mods, nil
 }
 
 func writeImportedModule(generatedDir, projectName string, module checker.Module, written map[string]struct{}) error {
@@ -83,7 +93,11 @@ func writeImportedModule(generatedDir, projectName string, module checker.Module
 	if err := os.WriteFile(outputPath, source, 0o644); err != nil {
 		return err
 	}
-	for _, mod := range requiredImportedModules(module.Program(), projectName) {
+	mods, err := requiredImportedModules(module, projectName)
+	if err != nil {
+		return err
+	}
+	for _, mod := range mods {
 		if err := writeImportedModule(generatedDir, projectName, mod, written); err != nil {
 			return err
 		}
